Use integer zoom levels; fix markdown links

+++ src/main/scala/org/leaflet/LMap.scala
@@ -47,7 +47,7 @@ class LMap(id: String, options: UndefOr[LMapOptions]) extends LEventEmitter {
   //Methods for Getting Map State
   def getCenter(): LLatLng = js.native
 
-  def getZoom(): Double = js.native
+  def getZoom(): Int = js.native
 
   def getMinZoom(): Double = js.native
